+++ src/array/ft_array_destroy.c
@@ -6,7 +6,7 @@
 /*   By: abrabant <abrabant@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/21 14:54:20 by abrabant          #+#    #+#             */
-/*   Updated: 2021/03/21 17:58:15 by abrabant         ###   ########.fr       */
+/*   Updated: 2021/03/27 13:17:07 by abrabant         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 void	ft_array_destroy(t_array_int *a, void (*fn)(void *el))
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (fn != NULL)
